Moving Box Blur From ETL To Synfig-Core
Moving Boxblur from ETL to Synfig Core
