+++ ETL/ETL/CMakeLists.txt
@@ -13,7 +13,6 @@ set(ETL_HEADERS
         "${CMAKE_CURRENT_LIST_DIR}/surface"
         "${CMAKE_CURRENT_LIST_DIR}/clock"
         "${CMAKE_CURRENT_LIST_DIR}/stringf"
-        "${CMAKE_CURRENT_LIST_DIR}/boxblur"
         "${CMAKE_CURRENT_LIST_DIR}/smart_ptr"
         "${CMAKE_CURRENT_LIST_DIR}/ref_count"
         "${CMAKE_CURRENT_LIST_DIR}/pen"
@@ -27,7 +26,6 @@ set(ETL_HEADERS
         "${CMAKE_CURRENT_LIST_DIR}/_smart_ptr.h"
         "${CMAKE_CURRENT_LIST_DIR}/_pen.h"
         "${CMAKE_CURRENT_LIST_DIR}/_bezier_angle.h"
-        "${CMAKE_CURRENT_LIST_DIR}/_boxblur.h"
         "${CMAKE_CURRENT_LIST_DIR}/_angle.h"
         "${CMAKE_CURRENT_LIST_DIR}/etl_config.h"
         "${CMAKE_CURRENT_LIST_DIR}/_stringf.h"
